requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
backcall==0.2.0
cryptography==43.0.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
faiss-cpu==1.8.0.post1
fastapi==0.112.0
frozenlist==1.4.1
greenlet==3.0.3
h5py==3.11.0
hpack==4.0.0
ipython==8.12.3
jsonpatch==1.33
jsonpointer==2.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.25
langchain-text-splitters==0.2.2
langsmith==0.1.94
llama-cloud==0.0.11
llama-index==0.10.58
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.58
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.31
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llama_cpp_python==0.2.82
markdownify==0.13.1
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
ollama==0.3.1
openai==1.37.2
orjson==3.10.6
pandas==2.2.2
pdfminer.six==20240706
pickleshare==0.7.5
pipreqs==0.5.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdf==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
sse-starlette==2.1.3
starlette==0.37.2
starlette-context==0.3.6
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.7.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
typing-inspect==0.9.0
tzdata==2024.1
uvicorn==0.30.5
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
